refactor: move presentations tests to tests own tests folder
